django-rest-framework/employees/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import Employee
from .serializers import EmployeeSerializer
# Create your views here.


def index(request):
    return render(request, 'index.html', {})


class EmployeeViewSet(viewsets.ModelViewSet):
    queryset = Employee.objects.all()
    serializer_class = EmployeeSerializer
    # lookup_field = 'employee_id'
